# docs/openapi.json
{
  "components": {
...
    "responses": {
      "BadRequest": {
        "description": "Bad Request",
        "examples": [
          {
            "error": "Bad Request",
            "message": "Invalid input data.",
            "success": false
          }
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "message",
          "error"
        ],
        "type": "object"
      },
...
      "Forbidden": {
        "description": "Forbidden",
        "examples": [
          {
            "error": "Forbidden.",
            "message": "You do not have permission to access this resource.",
            "success": false
          }
        ],
        "properties": {
          "error": {
            "description": "Error details, if any",
            "example": "Forbidden.",
            "type": "string"
          },
          "message": {
            "description": "A message providing additional information",
            "example": "You do not have permission to access this resource.",
            "type": "string"
          },
          "success": {
            "description": "Indicates the request failed due to insufficient permissions",
            "example": false,
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "InternalServerError": {
        "description": "Internal Server Error",
        "examples": [
          {
            "error": "Internal Server Error.",
            "message": "An unexpected error occurred.",
            "success": false
          }
        ],
        "properties": {
          "error": {
            "description": "Error details, if any",
            "example": "Internal Server Error.",
            "type": "string"
          },
          "message": {
            "description": "A message providing additional information",
            "example": "An unexpected error occurred.",
            "type": "string"
          },
          "success": {
            "description": "Indicates the request failed due to a server error",
            "example": false,
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "NotFound": {
        "description": "Not Found",
        "examples": [
          {
            "error": "Not Found.",
            "message": "Resource not found.",
            "success": false
          }
        ],
        "properties": {
          "error": {
            "description": "Error details, if any",
            "example": "Not Found.",
            "type": "string"
          },
          "message": {
            "description": "A message providing additional information",
            "example": "Resource not found.",
            "type": "string"
          },
          "success": {
            "description": "Indicates the requested resource could not be found",
            "example": false,
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "examples": [
          {
            "error": "Unauthorized.",
            "message": "Unauthorized access.",
            "success": false
          }
        ],
        "properties": {
          "error": {
            "description": "Error details, if any",
            "example": "Unauthorized.",
            "type": "string"
          },
          "message": {
            "description": "A message providing additional information",
            "example": "Unauthorized access.",
            "type": "string"
          },
          "success": {
            "description": "Indicates the request failed due to authentication failure",
            "example": false,
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
    },
    "schemas": {
...
      "FileUploadRequest": {
        "description": "Request payload for uploading a file to a project.",
        "example": {
          "files": [
            "design_mockup.png"
          ]
        },
        "properties": {
          "files": {
            "description": "List of files to be uploaded (multipart/form-data).",
            "items": {
              "format": "binary",
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "files"
        ],
        "title": "File Upload Request",
        "type": "object"
      },
...
      "ProjectFileResponse": {
        "description": "Response payload containing details of an uploaded project file.",
        "examples": [
          {
            "created_at": "2024-02-07T12:34:56Z",
            "file_name": "design_mockup.png",
            "file_size": 204800,
            "file_type": "png",
            "id": "123e4567-e89b-12d3-a456-426614174000",
            "project_id": "987e6543-e21b-45c3-b123-789456123000",
            "uploaded_by": "john.doe@example.com",
            "url": "https://cdn.example.com/uploads/projects/987e6543/design_mockup.png"
          }
        ],
        "properties": {
          "created_at": {
            "description": "Timestamp when the file was uploaded.",
            "example": "2024-02-07T12:34:56Z",
            "format": "date-time",
            "type": "string"
          },
          "file_name": {
            "description": "Name of the uploaded file.",
            "example": "design_mockup.png",
            "type": "string"
          },
          "file_size": {
            "description": "File size in bytes.",
            "example": 204800,
            "format": "int64",
            "type": "integer"
          },
          "file_type": {
            "description": "MIME type of the file.",
            "example": "image/png",
            "type": "string"
          },
          "id": {
            "description": "Unique file ID.",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid",
            "type": "string"
          },
          "project_id": {
            "description": "Associated project ID.",
            "example": "987e6543-e21b-45c3-b123-789456123000",
            "format": "uuid",
            "type": "string"
          },
          "uploaded_by": {
            "description": "Email or identifier of the user who uploaded the file.",
            "example": "john.doe@example.com",
            "type": "string"
          },
          "url": {
            "description": "Public URL to access the file.",
            "example": "https://cdn.example.com/uploads/projects/987e6543/design_mockup.png",
            "format": "uri",
            "type": "string"
          }
        },
        "title": "Project File Response",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "description": "JWT Token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "support@projectmanagement.com",
      "name": "Project Management API Team",
      "url": "http://projectmanagement.com"
    },
    "description": "API for managing clients, projects, and issues to demonstrate the features of OpenAPI 3.1 specification.",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "termsOfService": "http://example.com/terms/",
    "title": "Project Management API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/project/{project_id}/files": {
      "get": {
        "description": "Fetch a list of files uploaded to the specified project.",
        "operationId": "listProjectFiles",
        "parameters": [
          {
            "description": "Unique identifier of the project.",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "items": {
                        "$ref": "#/components/schemas/ProjectFileResponse"
                      },
                      "type": "array"
                    },
                    "message": {
                      "example": "Project files retrieved successfully.",
                      "type": "string"
                    },
                    "success": {
                      "example": true,
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Successfully retrieved project files."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/BadRequest"
                }
              }
            },
            "description": "Bad request due to invalid project ID."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/Unauthorized"
                }
              }
            },
            "description": "Unauthorized request."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/Forbidden"
                }
              }
            },
            "description": "Forbidden request."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/NotFound"
                }
              }
            },
            "description": "Project or files not found."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/InternalServerError"
                }
              }
            },
            "description": "Internal server error."
          }
        },
        "summary": "Retrieve a list of project files",
        "tags": [
          "ProjectFiles"
        ]
      },
      "post": {
        "description": "Upload one or more files to the specified project.",
        "operationId": "uploadProjectFiles",
        "parameters": [
          {
            "description": "Unique identifier of the project.",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/FileUploadRequest"
              }
            }
          },
          "description": "The request body for uploading a file, using multipart/form-data to include file metadata and contents.",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "items": {
                        "$ref": "#/components/schemas/ProjectFileResponse"
                      },
                      "type": "array"
                    },
                    "message": {
                      "example": "Files uploaded successfully.",
                      "type": "string"
                    },
                    "success": {
                      "example": true,
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Files uploaded successfully."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/BadRequest"
                }
              }
            },
            "description": "Bad request due to invalid project ID or file format."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/Unauthorized"
                }
              }
            },
            "description": "Unauthorized request."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/Forbidden"
                }
              }
            },
            "description": "Forbidden request."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/NotFound"
                }
              }
            },
            "description": "Project not found."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/InternalServerError"
                }
              }
            },
            "description": "Internal server error."
          }
        },
        "summary": "Upload multiple files to a project",
        "tags": [
          "ProjectFiles"
        ]
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Test server",
      "url": "https://test.projectmanagement.com/api/v1"
    },
    {
      "description": "Development server",
      "url": "https://dev.projectmanagement.com/api/v1"
    },
    {
      "description": "Production server",
      "url": "https://api.projectmanagement.com/api/v1"
    }
  ],
...
}